pyproject.toml
[project]
name = "polymarket-pipeline"
version = "0.1.0"
description = "Incremental batch pipeline ingesting and aggregating Polymarket orderbook data into Google BigQuery."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.5.0",
    "google-auth>=2.49.1",
    "google-cloud-bigquery>=3.40.1",
    "google-cloud-storage>=3.10.1",
    "pandas>=3.0.1",
    "pandas-gbq>=0.34.1",
    "pyarrow>=23.0.1",
    "requests>=2.32.5",
    "urllib3>=2.6.3",
]
